move CPU weight pack functions to layers/amx_utils.py
